refactor(project): convert github installations fetch as composable
